added new class HasExamples.cs (extracted tests from ApiExamples.cs)
